tools/requirements.txt
# NEURON Visual Fidelity Tools - Python Dependencies
# Install with: pip install -r requirements.txt

pillow>=9.0.0
opencv-python>=4.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
scikit-image>=0.19.0
